% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{ATE}
\alias{ATE}
\title{Compute Average Treatment Effect}
\usage{
ATE(FLAME_object)
}
\arguments{
\item{FLAME_object}{object returned by applying FLAME matching algorithm
(\code{\link{FLAME_bit}} or \code{\link{FLAME_PostgreSQL}} or
\code{\link{FLAME_SQLite}})}
}
\value{
Average Treatment Effect (ATE)
}
\description{
\code{ATE} computes average treatment effect for all matched groups.
}
