% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{CATE}
\alias{CATE}
\title{Compute Conditional Average Treatment Effect}
\usage{
CATE(FLAME_object, num_covs)
}
\arguments{
\item{FLAME_object}{object returned by applying FLAME matching algorithm
(\code{\link{FLAME_bit}} or \code{\link{FLAME_PostgreSQL}} or
\code{\link{FLAME_SQLite}})}

\item{num_covs}{number of covariates}
}
\value{
Conditional Average Treatment Effect (CATE)
}
\description{
\code{CATE} computes conditional average treatment effect given number of
covariates.
}
