% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Average_CATE.R
\name{CATE_AVG}
\alias{CATE_AVG}
\title{compute average CATE}
\usage{
CATE_AVG(FLAME_object, num_covs)
}
\arguments{
\item{FLAME_object}{object returned by applying the FLAME algorithm
(\code{\link{FLAME_bit}} or \code{\link{FLAME_PostgreSQL}} or
\code{\link{FLAME_SQLite}})}

\item{num_covs}{number of covariates used for matching}
}
\value{
Average CATE
}
\description{
\code{CATE_AVG} computes average CATE given number of covariates used for matching.
}
