% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_plot.R
\name{CATE_plot}
\alias{CATE_plot}
\title{summarize CATE of all matched groups by boxplot}
\usage{
CATE_plot(FLAME_object, num_covs)
}
\arguments{
\item{FLAME_object}{object returned by applying the FLAME algorithm
(\code{\link{FLAME_bit}} or \code{\link{FLAME_PostgreSQL}} or
\code{\link{FLAME_SQLite}})}

\item{num_covs}{number of covariates used for matching}
}
\value{
boxplot
}
\description{
Given number of covariates used for matching, \code{CATE_plot} visualizes
CATE of all matched groups by boxplot
}
