% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pairwise_integrals.R
\name{compute_pairwise_integrals}
\alias{compute_pairwise_integrals}
\title{Integral of the pairwise product of functions}
\usage{
compute_pairwise_integrals(matrix, T_grid)
}
\arguments{
\item{matrix}{matrix. \code{m} \eqn{\times} \code{J} matrix containing in the column
\eqn{j} the pointwise representation of the function \eqn{f_j} on the \code{T_grid} domain.}

\item{T_grid}{vector. \code{m}-length vector of the time domain}
}
\value{
matrix \code{J} \eqn{\times} \code{J} containing in each element \eqn{(i,j)}
the integral of \eqn{f_i \cdot f_j} on the domain defined by \code{T_grid}.
}
\description{
Given a set of functions, represented as their pointwise evaluation on a time grid
and stored in the columns of a matrix, it computes the integral of the pairwise
product of the functions
}
\examples{
data(SobolevKernel)
T_domain <- seq(0, 1, length = 50)
pairwise_derivartives <- compute_pairwise_integrals(derivatives, T_domain)

}

