% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMaterniso.R
\name{covMaterniso}
\alias{covMaterniso}
\title{Compute the Covariance of a Matern process}
\usage{
covMaterniso(nu, rho, sigma, x)
}
\arguments{
\item{nu}{scalar. Value s.t. 2\code{nu} is integer. It is a smoothing
paramter \eqn{\nu} of the covariance: the larger \code{nu}, the smoother the process.}

\item{rho}{scalar. Non negative range parameter \eqn{\rho} of the covariance.}

\item{sigma}{scalar. Non negative standard deviation parameter \eqn{s} of the covariance.}

\item{x}{vector. Vector containing the
abscissa points where the Covariance is defined.}
}
\value{
\code{d} \eqn{\times}\code{d} covariance matrix.
}
\description{
Definition of the Matern Covariance function, if \eqn{2 \nu} is integer. For
\eqn{\nu =} \code{0.5} the function is also known as the exponential covariance  or the
Ornstein-Uhlenbeck covariance in one dimension. In general the explicit formula
for the Matern Covariance for one dimensional processes, with \eqn{\nu} s.t. \eqn{\nu = p + 0.5} and \eqn{p} integer is:
\deqn{
k(d) = s^2  \frac{p!}{(2p)!} exp\left(-\frac{\sqrt{2 \nu} d}{\rho}\right) \sum_{i = 0}^p \frac{(p+i)!}{(p-i)! i!} \left( \frac{\sqrt{8 \nu} d}{\rho} \right)^{p-i}
}
with \eqn{d} the distance between two points \eqn{x} and \eqn{y} in R.
}
\examples{
# Defintion of a Gaussian process
# with Matern Covariance

# Time domain of the Gaussian Process
M <- 50
T_domain <- seq(0, 1, length = M)

# paramteters of the Matern Covariance
nu_alpha <- 2.5
range <- 1/4
variance <- 1

# mean of the process
mu_alpha <- rep(0,M)

# covariance structure
Sig_alpha <- covMaterniso(nu_alpha, rho = range, sigma = sqrt(variance), T_domain)

# definition of the process

# alpha <- mvrnorm(mu=mu_alpha, Sigma=Sig_alpha, n=1) # if MASS is inslalled

}

