% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_beta.R
\name{estimation_beta}
\alias{estimation_beta}
\title{FLAME for the Function-on-Scalar regression coefficients.}
\usage{
estimation_beta(X, Y, eigenval, NoI, thres, number_non_zeros, ratio_lambda,
  number_lambda, proportion_training_set, verbose = FALSE)
}
\arguments{
\item{X}{matrix. \code{N} \eqn{\times} \code{I} design matrix. It has
standardized columns.}

\item{Y}{matrix. \code{J} \eqn{\times} \code{N} matrix containing the
coefficients of the projection of the \code{N} response functions on the kernel
basis. In particular column \eqn{n} contains the basis projection of \eqn{y_n}.}

\item{eigenval}{vector. \code{J}-length vector containing the eigenvalues
of the kernel.}

\item{NoI}{scalar. integer, maximum number of iterations in the
Coordinate-Descent loop.}

\item{thres}{scalar. tolerance on the K-norm of the increment of the estimation
to stop the Coordinate-Descent loop}

\item{number_non_zeros}{scalar. integer,
threshold on the number of non zeros parameters to be detected. It is the
kill switch parameter presented in the Vignette.}

\item{ratio_lambda}{scalar. ratio to compute the minimum value of lambda. The
maximum \eqn{\lambda_{\max}} is computed as the minimum value which makes all the coefficients
equal to zero. The minimum is the product \code{ratio_lambda}\eqn{\times \lambda_{\max}}.}

\item{number_lambda}{scalar. integer, length of the grid for the
\eqn{\lambda} parameter.}

\item{proportion_training_set}{scalar. value in (0,1), the
proportion for the training set for the Cross Validation.}

\item{verbose}{bool. If \code{TRUE} the progression of the algorithm in the
adaptive and non adaptive step is shown. If \code{FALSE} no output is shown.
Default is \code{FALSE}.}
}
\value{
list containing:
 \itemize{
 \item \code{beta} matrix. \code{J} \eqn{\times} \code{I} matrix of the final estimated
 coefficients in the kernel basis.
 \item \code{beta_no_adaptive}  matrix \code{J} \eqn{\times} \code{I} matrix of
 the coefficients estimated at the end of the non-adaptive step.
 \item \code{predictors} vector of the indices of the non-zero estimated predictors.
 \item \code{predictors_no_adaptive}  vector of the indices of the non-zeros
 predictor estimated after the non-adaptive step.}
}
\description{
This function computes  FLAME for the Function-on-Scalar
regression problem,
returning the significant parameters and
the estimated coefficients in the kernel basis.
}
\examples{
\dontrun{
data(simulation)
data(SobolevKernel)
I0 <- dim(B_true)[2]
time <- proc.time()
FLAME_estimation <- estimation_beta(X = X,
                                 Y = Y_matrix,
                                 eigenval = eigenval,
                                 NoI = 10,
                                 thres = 0.1,
                                 number_non_zeros = I0*2,
                                 ratio_lambda = 0.01,
                                 number_lambda = 100,
                                 proportion_training_set = 0.75,
                                 verbose = FALSE)
duration <- proc.time()-time
duration
names(FLAME_estimation)
}

}

