% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAME.R
\name{FLAME}
\alias{FLAME}
\title{Global FLAME method: from the definition of the kernel to the estimation}
\usage{
FLAME(Y, X, type_kernel = "sobolev", param_kernel = 8, thres_eigen = 0.99,
  period_kernel = NULL, NoI = 10, thres_CD = 0.01,
  number_non_zeros = NULL, ratio_lambda = 0.01, number_lambda = 100,
  proportion_training_set = 0.75, verbose = FALSE)
}
\arguments{
\item{Y}{fd object or list. \code{N} functional responses of the
Function-on-Scalar regression probelm. It can be an \link[=fda]{fd} object or
a list with 2 elements: \code{data} and \code{time_domain}. \code{time_domain}
is the \code{m}-length vector of the absicissa grid of the functions and \code{data}
is the \code{N} \eqn{\times} \code{m} matrix of the point-wise evaluation of
the response functions.}

\item{X}{matrix. \code{N} \eqn{\times} \code{I} design matrix. It has
standardized columns.}

\item{type_kernel}{string. Four possible choices are implemented.  \code{gaussian},
\code{exponential}, \code{sobolev} or \code{periodic}. For other
kernels, define manually the eigenfunctions and eigenvectors and then use
the \link[=estimation_beta]{estimation_beta} function. Defualt is \code{sobolev}.}

\item{param_kernel}{scalar. Value of the characteristic smoothing parameter of the kernel.
It is the \eqn{\sigma}
parameter of the Gaussian and the Exponential kernel, as introduced
 in \link[kernlab]{rbfdot} and \link[kernlab]{laplacedot} functions;
 the \eqn{\sigma} parameter of  the Sobolev
kernel as in the \link[=sobolev_kernel_generation]{sobolev_kernel_generation} function or
the \eqn{\sigma} paramter of the periodic kernel of the
\link[=generation_kernel_periodic]{generation_kernel_periodic} function. Default is
\code{8}.}

\item{thres_eigen}{scalar. Threshold to identify the significant
eigenvalues of the kernel. The number of significant eigennvalues \code{J} is the
minimum \eqn{J} s.t.
\deqn{
\sum_{j = 1}^{J} \theta_j \geq \textrm{thres\_eigen} \sum_{j = 1}^{\infty} \theta_j.
}
Default is \code{0.99}.}

\item{period_kernel}{scalar. Period of the kernel. In case
 of \code{type_kernel = "periodic"}, it is a mandatory parameter with no default.
If other types of kernel are chosen, it is ignored.}

\item{NoI}{scalar. integer, maximum number of iterations in the
Coordinate-Descent loop. Default is \code{10}.}

\item{thres_CD}{scalar. tolerance in the increment of the K-norm of the estimation
to stop the Coordinate-Descent loop.
Default is \code{0.01}}

\item{number_non_zeros}{scalar. integer,
threshold on the number of non zeros parameters to be detected. It is the
kill switch parameter. See the Vignette for further details. Default is \code{NULL}
meaning that no kill switch paramter is imposed.}

\item{ratio_lambda}{scalar. ratio to compute the minimum value of lambda. The
maximum \eqn{\lambda} (\eqn{\lambda_{\textrm{max}}}) is computed as the minimum value which makes all the coefficients
equal to zero. And the minimum is the product \code{ratio_lambda}\eqn{\times \lambda_{\max}}. Default is \code{0.01}.}

\item{number_lambda}{scalar. integer, length of the grid for the
\eqn{\lambda} parameter. Default is \code{100}.}

\item{proportion_training_set}{scalar. value in (0,1), the
proportion for the training set for the Cross Validation.
Defualt is \code{0.75}.}

\item{verbose}{bool. If \code{TRUE} the progression of the algorithm in the
adaptive and non adaptive step is shown. If \code{FALSE} no output is shown.
Default is \code{FALSE}.}
}
\value{
list containing:
 \itemize{
 \item \code{beta} fd object or list. \code{I} functional coefficients
 estimated by FLAME. If \code{Y} is an \link[=fda]{fd} object. then
 also \code{beta} is  an \link[=fda]{fd} object, and \code{beta} is
 the projection of the estimations on a 10 elements cubic Bspline basis.
 If \code{Y} is a list, then also \code{beta} is a
 list with 2 elements: \code{data} and \code{time_domain}. \code{time_domain}
is the \code{m}-length domain grid, while \code{data} is \code{I} \eqn{\times} \code{m} matrix of the point-wise evaluation of
the estimated coefficients.
 \item \code{predictors} vector of the indices of
 the non-zero estimated predictors.}
}
\description{
It computes  FLAME  for the Function-on-Scalar
regression problem. From a set of functions -stored in a matrix or in
an \link[=fda]{fd} object- and a set of predictors, FLAME identifies the
set of meaningful predictors and their smooth representation in the
kernel space.
}
\examples{
\dontrun{
data(simulation)
data(SobolevKernel)
time <- proc.time()
FLAME_estimation <- FLAME()
duration <- proc.time()-time
duration
names(FLAME_estimation)
}

}

