% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAME_summary.R
\name{FLAME_summary}
\alias{FLAME_summary}
\title{Summary}
\usage{
FLAME_summary(FLAME_object)
}
\arguments{
\item{FLAME_object}{object returned by applying the FLAME algorithm
(\code{\link{FLAME_bit}}, \code{\link{FLAME_PostgreSQL}}, or
\code{\link{FLAME_SQLite}})}
}
\value{
(1) Number of units matched (2) Average treatment effect
(3) Summary of conditional average treatment effect (CATE) in boxplot
}
\description{
\code{FLAME_summary} provides brief summary of FLAME implementation.
}
