% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation_kernel.R
\name{generation_kernel}
\alias{generation_kernel}
\title{Definition of the eigenfunctions and eigenvalues of the kernel}
\usage{
generation_kernel(type = "sobolev", parameter = NULL, domain,
  thres = 0.99, return.derivatives = FALSE)
}
\arguments{
\item{type}{string. Type of kernel. Three possible choices implemented:
\code{gaussian},  \code{exponential} and \code{sobolev}. For the other
types of kernel, define manually the eigenfunctions and eigenvectors
of the kernel.}

\item{parameter}{scalar. Value of the characteristic parameter of the kernel.
It is the \eqn{\sigma}
parameter of the Gaussian and the Exponential kernel, as introduced in \code{kernlab};
and the \eqn{\sigma} parameter of  the Sobolev
kernel as in the \link[=sobolev_kernel_generation]{sobolev_kernel_generation} function.}

\item{domain}{vector. \code{m}-length vector of the
abscissa grid  for the definition of the kernel.}

\item{thres}{scalar. Threshold to identify the significant
eigenvalues of the kernel.  The number of significant eigennvalues \code{J} is
the minimum \eqn{J} s.t.
\deqn{
\sum_{j = 1}^J \theta_j \geq \textrm{thres} \sum_{j = 1}^{\infty} \theta_j.
} Default is 0.99.}

\item{return.derivatives}{bool. If \code{TRUE} the function returns also
the matrix of the evaluation of
the derivatives of the eigenfunctions on the time domain.
 Default is \code{FALSE}.}
}
\value{
list containing
 \itemize{
 \item \code{eigenvect} \code{m} \eqn{\times} \code{J} matrix of
 the eigenfunctions of the kernel evaluated on the \code{domain}.
 \item \code{eigenval} \code{J}-length  vector of the
 eigenvalues of the kernel
 \item \code{derivatives}. if \code{return.derivatives = TRUE}.
 \code{derivatives} is the (\code{m-1}) \eqn{\times} \code{J} matrix of the derivatives of
 the eigenfunctions evaluated on the time domain.
}
}
\description{
Given a particular type of kernel, to be chosen among (\code{gaussian},
\code{exponential} and \code{sobolev}), it returns the
evaluation of the eigenfunctions of the kernel on the grid \code{domain}
and the correspondent eigenvalues.
}
\details{
Here the list of the kernel defined in this function
\itemize{
\item \code{gaussian}
\deqn{
k(x, x') = \exp(-\sigma \| x- x'\|^2)
}
\item \code{exponential}
\deqn{
k(x, x') = \exp(-\sigma \| x- x'\|)
}
\item \code{sobolev}, the kernel associated to the norm in the \eqn{H^1} space
\deqn{
\| f \|^2 = \int_{D} f(t)^2 dt + \frac{1}{\sigma} \int_{D} f'(t)^2 dt
}
where \eqn{D} is the one-dimensional \code{domain} and \eqn{f'} is the first derivative of the function.
}
}
\examples{
# definition of the kernel
type_kernel <-  'sobolev'
param_kernel <-  8
T_domain <- seq(0, 1, length = 50)
kernel_here <- generation_kernel ( type = type_kernel,
                                   parameter = param_kernel,
                                   domain = T_domain,
                                   thres = 0.99,
                                   return.derivatives = TRUE)

eigenvalues <- kernel_here$eigenval
eigenvectors <- kernel_here$eigenvect
der <- kernel_here$derivatives

}

