% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation_kernel_periodic.R
\name{generation_kernel_periodic}
\alias{generation_kernel_periodic}
\title{Definition of the eigenfunctions and eigenvalues of the periodic kernel}
\usage{
generation_kernel_periodic(period = NULL, parameter = NULL, domain,
  thres = 0.99, return.derivatives = FALSE)
}
\arguments{
\item{period}{scalar. Period of the kernel.}

\item{parameter}{scalar. Parameter to tune the smoothness level of the kernel.
The \eqn{\sigma} parameter introduced in Details.}

\item{domain}{vector. \code{m}-length vector for the abscissa grid
of the kernel.}

\item{thres}{scalar. Threshold for the identification of the significant
eigenvalues of the kernel. The number of significant eigennvalues \code{J} is
the minimum \eqn{J} s.t.
\deqn{
\sum_{j = 1}^J \theta_j \geq \textrm{thres} \sum_{j = 1}^{\infty} \theta_j.
} Default is 0.99.}

\item{return.derivatives}{bool. If \code{TRUE} the function returns the matrix of
the derivatives of the selected eigenfunctions evaluated on the time domain.
 Default is \code{FALSE}.}
}
\value{
list containing
 \itemize{
 \item \code{eigenvect} \code{m} \eqn{\times} \code{J} matrix of
 the eigenfunctions of the kernel evaluated on the \code{domain}.
 \item \code{eigenval} \code{J}-length  vector of the
 eigenvalues of the kernel
 \item \code{derivatives}. if \code{return.derivatives = TRUE}.
 \code{m-1} \eqn{\times} \code{J} matrix of the derivatives of
 the eigenfunctions.
}
}
\description{
Given the period \code{p} and the smoothing parameter \eqn{\sigma}, it returns the
evaluation of the eigenfunctions of the kernel on the grid \code{domain}
and the correspondent eigenvalues.
}
\details{
The periodic kernel of period \eqn{p} defined in this function is
\deqn{
K(x, y) = \sigma^2 \exp \left\{  -2/\sigma \sin^2\left(\frac{\pi | x - y|}{p} \right)\right\}.
}
where \eqn{\sigma} is the smoothing parameter.
}
\examples{
param_kernel <- 8
T_domain <- seq(0, 1, length = 50)
kernel_here <- generation_kernel_periodic(period = 1/2,
                                          parameter = param_kernel,
                                          domain = T_domain,
                                          thres = 1-10^{-16},
                                          return.derivatives = TRUE)
names(kernel_here)

}

