% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-c-enrichment_types.R
\name{getEnrichmentTypes}
\alias{getEnrichmentTypes}
\title{Get All Registered Enrichment Types}
\usage{
getEnrichmentTypes()
}
\value{
Character vector of registered type IDs
}
\description{
Get All Registered Enrichment Types
}
