% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-helpers.R
\name{getExamplesPath}
\alias{getExamplesPath}
\title{Get path to examples data}
\usage{
getExamplesPath(filename)
}
\arguments{
\item{filename}{The example data filename}
}
\value{
Full path to the file
}
\description{
Get path to examples data
}
\keyword{internal}
