% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-helpers.R
\name{getExtdataPath}
\alias{getExtdataPath}
\title{Get path to a data file in extdata}
\usage{
getExtdataPath(...)
}
\arguments{
\item{...}{Path components relative to extdata/}
}
\value{
Full path to the file
}
\description{
Get path to a data file in extdata
}
\keyword{internal}
