% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-helpers.R
\name{getFlameRoot}
\alias{getFlameRoot}
\title{Get the package root directory}
\usage{
getFlameRoot()
}
\value{
Character string with the path to package root
}
\description{
Returns the root directory of the flame package.
Works both during development and after installation.
}
\keyword{internal}
