% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-helpers.R
\name{getOrganismsPath}
\alias{getOrganismsPath}
\title{Get path to organisms data}
\usage{
getOrganismsPath(filename)
}
\arguments{
\item{filename}{The organism data filename}
}
\value{
Full path to the file
}
\description{
Get path to organisms data
}
\keyword{internal}
