% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-c-enrichment_types.R
\name{getValidDatasources}
\alias{getValidDatasources}
\title{Get Valid Datasources for Enrichment Type}
\usage{
getValidDatasources(type)
}
\arguments{
\item{type}{Character string: "functional", "literature", etc.}
}
\value{
Character vector of valid datasource names
}
\description{
Returns the list of valid datasources for a specific enrichment type.
This is used by plot datasource pickers to show appropriate options.
}
\examples{
getValidDatasources("functional")  # c("GO:BP", "GO:MF", "KEGG", ...)
getValidDatasources("literature")  # c("PUBMED")
}
