% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_matrix_H.R
\name{norm_matrix_H}
\alias{norm_matrix_H}
\title{Computation of the H-norm of a set of functions}
\usage{
norm_matrix_H(M)
}
\arguments{
\item{M}{matrix. \code{J} \eqn{\times} \code{N} matrix containing in the column
\eqn{n} the coefficients of the projection of a function \eqn{y_n} on the
\code{J}-dimensional basis.}
}
\value{
vector of length \code{N} containing the H-norm of the functions
}
\description{
It computes the H norm of a set of functions represented as their projection on the
basis of the space H.
}
\examples{
data(simulation)
length(norm_matrix_H(Y_matrix))
sum(norm_matrix_H(Y_matrix))

}

