% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_matrix_H_derivatives.R
\name{norm_matrix_H_derivatives}
\alias{norm_matrix_H_derivatives}
\title{Computation of the H-norm of the derivatives of a set of functions}
\usage{
norm_matrix_H_derivatives(M, pairwise_derivatives)
}
\arguments{
\item{M}{matrix. \code{J} \eqn{\times} \code{N} matrix containing in the column
\eqn{n} the coefficients of the projection of a function \eqn{y_n} on the
\code{J}-dimensional basis functions.}

\item{pairwise_derivatives}{matrix. \code{J} \eqn{\times} \code{J} matrix
containing  the integral of the pairwise product of the basis functions. It can
be computed with the \link[=compute_pairwise_integrals]{compute_pairwise_integrals}
function.}
}
\value{
vector of length \code{N} containing the H-norm of the derivatives of
the functions
}
\description{
It computes the H norm of the derivatives of a set of functions
represented as their projection on a
basis of the space H.
}
\examples{
data(SobolevKernel)
T_domain <- seq(0, 1, length = 50)
pairwise_derivartives <- compute_pairwise_integrals(derivatives, T_domain)
sum(norm_matrix_H_derivatives(t(Y_matrix), pairwise_derivartives))
}

