% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_basis.R
\name{projection_basis}
\alias{projection_basis}
\title{Projection on the kernel basis}
\usage{
projection_basis(y, eigenvect, M_integ)
}
\arguments{
\item{y}{matrix. \code{N} \eqn{\times} \code{m} matrix containing in
row \eqn{n} the point-wise evaluation of the function \eqn{y_n}.}

\item{eigenvect}{matrix. \code{m} \eqn{\times} \code{J} matrix
containing in  column \eqn{j} the point-wise evaluation of an eigenfunction
\eqn{v_j} of the kernel evaluated on an equispaced time domain grid.}

\item{M_integ}{scalar, integer. Number of points of the domain \eqn{D}
of the functions (\code{m}) divided by the width of the domain \eqn{|D|}.
Constant value used to compute the integrals.}
}
\value{
\code{J} \eqn{\times} \code{N} matrix (\eqn{Y}) containing in position
\eqn{(j,n)} the integral of the product
of the eigenfunction \eqn{v_j} with the function \eqn{y_n}, i.e. the coefficient
correspondent to \eqn{v_j} of the projection of \eqn{y_n} on the kernel basis.
\deqn{
Y[i,j] = \int_{D} y_n(t)  v_j(t) dt
}
}
\description{
It computes the  projection of a set of
 functions on the kernel basis defined by a set of eigenvalues and
eigenfunctions
}
\examples{
data(simulation)
data(SobolevKernel)
summary(T_domain)
M_integ <- length(T_domain)/diff(range(T_domain))
projection_basis(Y_full, eigenvect, M_integ ) # projection on the J dimensional
# basis of the Y functions.
}

