% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_domain.R
\name{projection_domain}
\alias{projection_domain}
\title{Representation on the time domain of a function defined by the coefficients of the kernel basis}
\usage{
projection_domain(y, eigenfun)
}
\arguments{
\item{y}{matrix. \code{J} \eqn{\times} \code{N} matrix containing in
column \eqn{n} the coefficients of the projection of the function \eqn{y_n}
on the \code{J} eigenfunctions of the kernel.}

\item{eigenfun}{matrix. \code{m} \eqn{\times} \code{J} matrix containing
in each column the
point-wise evaluation of the eigenfunctions on the
kernel}
}
\value{
\code{N} \eqn{\times} \code{m} matrix containing in the row \eqn{n} the pointwise
evaluation of the function \eqn{y_n} on the domain \eqn{D} of length \code{m}.
}
\description{
It computes the pointwise evaluation of a function (or a set of
functions) on the time domain, given its (their)  projection on the kernel basis.
}
\examples{
data(SobolevKernel)
data(simulation)
projection_domain(Y_matrix, eigenvect) # projection of the data
# on the time domain seq(0, 1, length = 50)

}

