% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobolev_kernel_generation.R
\name{sobolev_kernel_generation}
\alias{sobolev_kernel_generation}
\title{Definition of the kernel for the Sobolev Space}
\usage{
sobolev_kernel_generation(a, b, m, sigma, plot.eigen = FALSE)
}
\arguments{
\item{a}{scalar. left end point of the domain.}

\item{b}{scalar. right end point of the domain.}

\item{m}{scalar, integer. number of points of the interval domain.}

\item{sigma}{scalar. weight \eqn{\sigma} associated to the derivative in
 the norm associated to the kernel. See Details
for the explicit definition of the norm.}

\item{plot.eigen}{bool. if \code{TRUE} the cumulative sum of the
eigenvalues of the kernel is plotted. Default is \code{FALSE}.}
}
\value{
list containing
\itemize{
\item \code{vectors} matrix. \code{m} \eqn{\times} \code{m} matrix containing
the eigenvectors of the kernel. Each column contains the evaluation of an
eigenfunction on the domain \code{seq(a, b, length = m)}.
\item \code{values} vector. \code{m} length vector containing the eigenvalues
of the kernel
}
}
\description{
Definition of the eigenfunctions and eigenvalues of the
kernel for the Sobolev Space \eqn{H^1((a,b))}.
}
\details{
The norm associated to the Sobolev kernel, dependent on the
smoothing parameter \eqn{\sigma} is
\deqn{
\|f\|^2 = \|f\|^2_{L^2} + 1/\sigma \|f^{\prime}\|^2_{L^2}
}
The function \link[=sobolev_kernel_generation]{sobolev_kernel_generation}
is implicitly called in the \link[=generation_kernel]{generation_kernel}
function when \code{type} parameter is \code{'sobolev'}. See the Vignette for
the explicit definition of the kernel.
}
\examples{
sobolev_kernel_generation(a = 0, b = 1, m = 100,
               sigma = 1, plot.eigen = FALSE)

}

