% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAME_summary.R
\name{summary}
\alias{summary}
\title{Summary}
\usage{
summary(FLAME_object)
}
\arguments{
\item{FLAME_object}{object returned by applying the FLAME algorithm
(\code{\link{FLAME_bit}}, \code{\link{FLAME_PostgreSQL}}, or
\code{\link{FLAME_SQLite}})}
}
\value{
(1) Number of units matched (2) Average treatment effect (ATE)
}
\description{
\code{FLAME_summary} provides brief summary of FLAME implementation.
}
\examples{
data(toy_data)
result <- FLAME::FLAME_bit(data = toy_data, holdout = toy_data)
FLAME::summary(result)
}
