% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plot.R
\name{summary_plot}
\alias{summary_plot}
\title{Summary Plot}
\usage{
summary_plot(FLAME_object)
}
\arguments{
\item{FLAME_object}{object returned by applying FLAME matching algorithm
(\code{\link{FLAME_bit}} or \code{\link{FLAME_PostgreSQL}} or
\code{\link{FLAME_SQLite}})}
}
\value{
summary plot of FLAME algorithm
}
\description{
\code{summary_plot} visualizes the covariate dropped, number of matched units, and its
conditional average treatment effect at each iteration.
}
