% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-c-enrichment_types.R
\name{supportsCombination}
\alias{supportsCombination}
\title{Check if Enrichment Type Supports Combination}
\usage{
supportsCombination(type)
}
\arguments{
\item{type}{Character string: "functional", "literature", etc.}
}
\value{
Logical TRUE if combination tab is supported
}
\description{
Check if Enrichment Type Supports Combination
}
