% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Toy Data}
\format{A data frame with 2190410  rows and 86 covariates:
\describe{
  \item{X1}{Covariate 1}
  \item{X2}{Covariate 2}
  ...
  \item{outcome}{Outcome Variable}
  \item{treated}{Treated = 1, Control = 0}
}}
\usage{
toy_data
}
\description{
This toy data has 100 observations/units with 20 categorical covariates.
}
\keyword{datasets}
