% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATE.R
\name{ATE}
\alias{ATE}
\title{Compute Average Treatment Effect}
\usage{
ATE(FLAME_object)
}
\arguments{
\item{FLAME_object}{object returned by applying the FLAME algorithm
(\code{\link{FLAME_bit}}, \code{\link{FLAME_PostgreSQL}}, or
\code{\link{FLAME_SQLite}})}
}
\value{
average treatment effect (ATE)
}
\description{
\code{ATE} computes average treatment effect for each given sub-population by
a weighted average of the estimated treatment effects in each matched group.
The weight is the number of units matched.
}
