% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE.R
\name{CATE}
\alias{CATE}
\title{detailed information of each matched group}
\usage{
CATE(FLAME_object, num_covs)
}
\arguments{
\item{FLAME_object}{object returned by applying the FLAME algorithm
(\code{\link{FLAME_bit}} or \code{\link{FLAME_PostgreSQL}} or
\code{\link{FLAME_SQLite}})}

\item{num_covs}{number of covariates used for matching}
}
\value{
data frame with all matched units, the size and CATE of each matched group
}
\description{
Given number of covariates used for matching, \code{CATE} returns all matched
units and their conditional average treatment effect. If no
units are matched, \code{CATE} will return nothing.
}
