% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Average_CATE.R
\name{CATE_AVG}
\alias{CATE_AVG}
\title{compute average CATE}
\usage{
CATE_AVG(CATE_object, num_covs)
}
\arguments{
\item{CATE_object}{object returned by applying \code{\link{CATE}} function}
}
\value{
(weighted) average CATE
}
\description{
\code{CATE_AVG} computes (weighted) average CATE given CATE_object
}
