% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_plot.R
\name{CATE_plot}
\alias{CATE_plot}
\title{summarize CATE of all matched groups by boxplot}
\usage{
CATE_plot(CATE_object)
}
\arguments{
\item{CATE_object}{object returned by applying \code{\link{CATE}} function}
}
\value{
boxplot
}
\description{
Given CATE_object, \code{CATE_plot} visualizes CATE of all matched groups by
boxplot
}
