% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_plot.R
\name{CATE_plot}
\alias{CATE_plot}
\title{Summarize CATEs of All Matched Groups by Boxplot}
\usage{
CATE_plot(CATE_object)
}
\arguments{
\item{CATE_object}{object returned by applying \code{\link{CATE}} function}
}
\value{
boxplot
}
\description{
Given CATE_object, \code{CATE_plot} visualizes CATEs of all matched groups by
boxplot
}
\examples{
data(toy_data)
result <- FLAME::FLAME_bit(data = toy_data, holdout = toy_data)
CATE_object <- FLAME::CATE(FLAME_object = result, cov_name = c("X1", "X2"), cov_val = c("2", "2"))
FLAME::CATE_plot(CATE_object)
}
