% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Generation.R
\name{Data_Generation}
\alias{Data_Generation}
\title{Generate Data}
\usage{
Data_Generation(num_control, num_treated, num_cov_dense, num_covs_unimportant)
}
\arguments{
\item{num_control}{Number of controlled samples}

\item{num_treated}{Number of treated samples}

\item{num_cov_dense}{Number of covariates}

\item{num_covs_unimportant}{Number of unimportant covariates}
}
\value{
Data Frame
}
\description{
Generate Data
}
\examples{
Data_Generation(10,10,10,0)
Data_Generation(20,20,10,10)
}
