% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAMEbit.R
\name{FLAME_bit}
\alias{FLAME_bit}
\title{FLAME: Bit Vector Method}
\usage{
FLAME_bit(data, holdout, num_covs, num_treated, num_control, covs_max_list,
  tradeoff)
}
\arguments{
\item{holdout}{Holdout Training Data}

\item{num_covs}{Number of covariates}

\item{num_treated}{Number of units in treated group}

\item{num_control}{Number of units in control group}

\item{covs_max_list}{List indicates each covariate is binary/ternary/...}

\item{tradeoff}{Tradeoff parameter to compute Match Quality}

\item{df}{Data Frame}
}
\value{
Data Frame
}
\description{
FLAME: Bit Vector Method
}
