% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAMEbit.R
\name{FLAME_bit}
\alias{FLAME_bit}
\title{FLAME: Bit Vector Method}
\usage{
FLAME_bit(df, holdout, covs, covs_max_list, num_treated, num_control)
}
\arguments{
\item{df}{Data Frame}

\item{holdout}{Holdout Training Data}

\item{covs}{List of covariates}

\item{covs_max_list}{List indicates each covariate is binary/ternary/...}

\item{num_treated}{Number of units in treated group}

\item{num_control}{Number of units in control group}
}
\value{
Data Frame
}
\description{
FLAME: Bit Vector Method
}
