% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAMEbit.R
\name{FLAME_bit}
\alias{FLAME_bit}
\title{Bit Vectors Implementation}
\usage{
FLAME_bit(data, holdout, num_covs, num_treated, num_control, covs_max_list,
  tradeoff)
}
\arguments{
\item{data}{Input data}

\item{holdout}{Holdout training data}

\item{num_covs}{Number of covariates}

\item{num_treated}{Number of units in treated group}

\item{num_control}{Number of units in control group}

\item{covs_max_list}{List indicates each covariate is binary/ternary/...}

\item{tradeoff}{Tradeoff parameter to compute Match Quality}
}
\value{
(1) List of covariates matched at each iteration (2) list of data frame
showing the size of matched group and its conditional average treatment effect (CATE)
}
\description{
\code{FLAME_bit} applies FLAME matching algorithm based on bit vectors implementation.
}
