% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAME_SQLite.R
\name{FLAME_SQLite}
\alias{FLAME_SQLite}
\title{FLAME: SQLite Database}
\usage{
FLAME_SQLite(db, data, holdout, num_covs, tradeoff)
}
\arguments{
\item{db}{Name of the Database Connection}

\item{data}{Data Frame}

\item{holdout}{Holdout Training Data}

\item{num_covs}{Number of Covariates}

\item{tradeoff}{Tradeoff Parameter to compute Match Quality}
}
\value{
List of covariates matched in each iteration, CATE for matched units
}
\description{
FLAME: SQLite Database
}
