% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAME_SQLite.R
\name{FLAME_SQLite}
\alias{FLAME_SQLite}
\title{SQLite database implementation}
\usage{
FLAME_SQLite(db, data, holdout, num_covs, tradeoff = 0.1,
  PE_function = NULL, model = NULL, ridge_reg = NULL, lasso_reg = NULL,
  tree_depth = NULL)
}
\arguments{
\item{db}{name of the connection to temporary database  (\strong{must name the
connection as db})}

\item{data}{input data}

\item{holdout}{holdout training data}

\item{num_covs}{number of covariates}

\item{tradeoff}{tradeoff parameter to compute Matching Quality (default = 0.1)}

\item{PE_function}{user defined function to compute predivtive error (optional)}

\item{model}{user defined model - Linear, Ridge, Lasso, or DecisionTree  (optional)}

\item{ridge_reg}{L2 regularization parameter if model = Ridge (optional)}

\item{lasso_reg}{L1 regularization parameter if model = Lasso (optional)}

\item{tree_depth}{maximum depth of decision tree if model = DecisionTree (optional)}
}
\value{
(1) List of covariates matched at each iteration (2) List of data
 frame showing matched groups, conditional average treatment effect (CATE),
 and the size of each matched group
}
\description{
\code{FLAME_SQLite} applies the FLAME algorithm based on SQLite.
\code{FLAME_SQLite} does not require external database installment. However,
user should connect to a temporary database with command
\code{dbConnect(SQLite(),"tempdb_name")} and name the connection as
\strong{db}.
}
