% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAMEdb.R
\name{FLAME_SQLite}
\alias{FLAME_SQLite}
\title{FLAME: Database Method}
\usage{
FLAME_SQLite(db, data, holdout, num_covs, tradeoff)
}
\arguments{
\item{db}{Name of Database}

\item{data}{Data Frame}

\item{holdout}{Holdout Training Data}

\item{num_covs}{Number of Covariates}

\item{tradeoff}{Tradeoff Parameter to compute Match Quality}
}
\value{
List of covariates matched in each iteration, CATE for matched units
}
\description{
FLAME: Database Method
}
